Move test to correct test class.
